remove precompile
